+++ rosys/geometry/__init__.py
@@ -1,6 +1,4 @@
-from .geo_point import Fixpoint, GeoPoint
-from .geo_pose import GeoPose
-from .geo_reference import GeoReference, current_geo_reference
+from .geo import Fixpoint, GeoPoint, GeoPose, GeoReference, current_geo_reference
 from .line import Line
 from .line_segment import LineSegment
 from .object3d import Object3d
