+++ rosys/hardware/__init__.py
@@ -7,7 +7,7 @@ from .communication import Communication, SerialCommunication, WebCommunication
 from .esp_pins import EspPins
 from .estop import EStop, EStopHardware, EStopSimulation
 from .expander import ExpanderHardware
-from .gnss import Gnss, GnssHardware, GnssMeasurement, GnssSimulation, LocalGnssPoseProvider
+from .gnss import Gnss, GnssHardware, GnssMeasurement, GnssSimulation
 from .imu import Imu, ImuHardware, ImuSimulation
 from .module import Module, ModuleHardware, ModuleSimulation
 from .robot import Robot, RobotHardware, RobotSimulation
@@ -30,7 +30,6 @@ __all__ = [
     'GnssHardware',
     'GnssMeasurement',
     'GnssSimulation',
-    'LocalGnssPoseProvider',
     'SerialCommunication',
     'WebCommunication',
     'EspPins',
